Handle refresh issues related to X-FRAME-OPTIONS
When both access and refresh tokens expire, Azure AD
edirects to a page with X-FRAME-OPTIONS set to DENY,
which does not  cause any javascript errors, but the
refresh token call fails and does not resume. This code
adds a timeout after which we check if the request for
a refresh was successful and if not, we clean up all
callbacks and pass them the error.